Update documentation.md
The 'default' text does not work and would be weird to support.
Searching in code wouldn't make sense because the client changed the text anyways?
Also if client leaves this field empty, you will render empty HTML elements. Which makes me even think this entire feature should be removed?

BennoThommo:
I've noticed the same issue before. Currently, it won't use a default value like that, and I think the reasoning is because when you save the field in the Backend without entering anything in the field, it saves the field as an empty value. The problem is, there's no way to tell if it's intentionally empty or not - for example, if you're offering this theme to a client to modify, they may intentionally not want a tagline to be used on a page, so they would leave the field blank.